+++ src/com/jordanbunke/tdsm_api/ast/expr/init/InitAssetChoiceNode.java
@@ -44,9 +44,7 @@ public final class InitAssetChoiceNode extends InitExprNode {
                 .toArray(ColorSelection[]::new);
 
         if (id.isEmpty()) {
-            ScriptErrorLog.fireError(
-                    ScriptErrorLog.Message.CUSTOM_RT,
-                    arguments.get(0).getPosition(),
+            ScriptErrorLog.runtimeError(arguments.get(0).getPosition(),
                     "Asset choice ID must be non-empty");
             return null;
         }